+++ src/ui/item.cpp
@@ -858,7 +858,7 @@ bool wex::item::validate() const
   {
     try
     {
-      std::regex r(get_value_as_string());
+      boost::regex r(get_value_as_string());
     }
     catch (std::exception&)
     {
@@ -873,7 +873,7 @@ bool wex::item::validate(const std::string& regex) const
 {
   try
   {
-    return std::regex_match(get_value_as_string(), std::regex(regex));
+    return boost::regex_match(get_value_as_string(), boost::regex(regex));
   }
   catch (const std::exception& e)
   {